Move MakeTx types to common
