+++ src/ripple/rippleTypes.ts
@@ -7,9 +7,9 @@ import {
   asOptional,
   asString
 } from 'cleaners'
-import { EdgeMetadata, EdgeTransaction, EdgeTxSwap } from 'edge-core-js/types'
+import { EdgeTransaction } from 'edge-core-js/types'
 
-import { asSafeCommonWalletInfo } from '../common/types'
+import { asSafeCommonWalletInfo, MakeTxParams } from '../common/types'
 
 export interface XrpNetworkInfo {
   rippledServers: string[]
@@ -62,29 +62,6 @@ export type RipplePrivateKeys = ReturnType<typeof asRipplePrivateKeys>
 export const asRipplePrivateKeys = asObject({
   rippleKey: asString
 })
-
-/**
- * Template for a future generalized makeTx type that can be used for all
- * chains
- */
-export type MakeTxParams =
-  | {
-      type: 'MakeTxDexSwap'
-      metadata?: EdgeMetadata
-      swapData?: EdgeTxSwap
-      fromTokenId?: string
-      fromNativeAmount: string
-      toTokenId?: string
-      toNativeAmount: string
-
-      /**
-       * UNIX time (seconds) to expire the DEX swap if it hasn't executed
-       */
-      expiration?: number
-    }
-  | {
-      type: 'MakeTxDummyType'
-    }
 
 export interface RippleOtherMethods {
   makeTx: (makeTxParams: MakeTxParams) => Promise<EdgeTransaction>
